feat: implement FeedContext for managing feed state and data fetching; refactor FeedPage to utilize context
