+++ constants/api.ts
@@ -1,0 +1,4 @@
+export const API_FEED = {
+  GET: "/api/v1/posts",
+
+};
